fixed up docs and start builds. shared components
