+++ helpers/example/components/brands.js
@@ -1,0 +1,15 @@
+import bomBrand from '@westpac/bom';
+import bsaBrand from '@westpac/bsa';
+import btfgBrand from '@westpac/btfg';
+import stgBrand from '@westpac/stg';
+import wbcBrand from '@westpac/wbc';
+import wbgBrand from '@westpac/wbg';
+
+export const BRANDS = {
+	BOM: bomBrand,
+	BSA: bsaBrand,
+	BTFG: btfgBrand,
+	STG: stgBrand,
+	WBC: wbcBrand,
+	WBG: wbgBrand,
+};
