+++ helpers/example/docs.js
@@ -1,358 +1,9 @@
-/** @jsx jsx */
-
-import { BrowserRouter as Router, Route, Switch, NavLink, useLocation } from 'react-router-dom';
-import React, { Fragment, useEffect, useState, StrictMode } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
 
-import { Global, css, jsx } from '@westpac/core';
-import { Container } from '@westpac/grid';
-
-import { GEL } from '@westpac/core';
-
-import bomBrand from '@westpac/bom';
-import bsaBrand from '@westpac/bsa';
-import btfgBrand from '@westpac/btfg';
-import stgBrand from '@westpac/stg';
-import wbcBrand from '@westpac/wbc';
-import wbgBrand from '@westpac/wbg';
-
-// ==============================
-// Get the data
-// ==============================
-
-const BRANDS = {
-	BOM: bomBrand,
-	BSA: bsaBrand,
-	BTFG: btfgBrand,
-	STG: stgBrand,
-	WBC: wbcBrand,
-	WBG: wbgBrand,
-};
-
-// ==============================
-// Compose the pieces
-// ==============================
-
-const App = ({ components, packageName, pkg }) => {
-	const [brand, setBrand] = useState('WBC');
-
-	return (
-		<Router>
-			<StrictMode>
-				<GEL brand={BRANDS[brand]}>
-					<div
-						css={{
-							alignItems: 'stretch',
-							display: 'flex',
-							minHeight: '100vh',
-						}}
-					>
-						<Global
-							styles={css`
-								code {
-									font-family: Monaco, monospace;
-									font-size: 0.85em;
-								}
-								p > code {
-									background-color: #ffebe6;
-									color: #bf2600;
-									border-radius: '0.3rem';
-									display: inline-block;
-									padding: '0.1rem 0.3rem';
-								}
-							`}
-						/>
-						<Switch>
-							<Route
-								exact
-								path="/"
-								render={route => (
-									<Fragment>
-										<Sidebar components={components} brand={brand} setBrand={setBrand} />
-										<Home {...route} />
-									</Fragment>
-								)}
-							/>
-							{components.map(({ slug, ...props }) => (
-								<Route
-									key={slug}
-									exact
-									path={`/${slug}`}
-									render={route => (
-										<Fragment>
-											<Sidebar
-												components={components}
-												brand={brand}
-												setBrand={setBrand}
-												parent={slug}
-											/>
-											<Page {...route} {...props} brand={BRANDS[brand]} />
-										</Fragment>
-									)}
-								/>
-							))}
-						</Switch>
-					</div>
-				</GEL>
-			</StrictMode>
-		</Router>
-	);
-};
-
-class Page extends React.Component {
-	state = { error: null, info: null };
-
-	componentDidCatch(error, info) {
-		this.setState({ error, info });
-	}
-
-	render() {
-		const { Module, filename, label, landing, ...rest } = this.props;
-		const { error, info } = this.state;
-		console.log(Module);
-
-		if (error) {
-			const errorLabel = `Error in ${filename}`;
-			console.error(errorLabel, error);
-
-			return (
-				<Article>
-					<Container>
-						<h1>{errorLabel}</h1>
-						<h2 css={{ color: '#BF2600' }}>
-							<code>{error.message}</code>
-						</h2>
-						<pre
-							css={{
-								backgroundColor: '#FFEBE6',
-								borderRadius: 4,
-								color: '#DE350B',
-								paddingBottom: '1em',
-							}}
-						>
-							<code>{info.componentStack}</code>
-						</pre>
-					</Container>
-				</Article>
-			);
-		}
-
-		if (landing) {
-			return <Home />;
-		}
-
-		return (
-			<Article>
-				<Container css={{ marginBottom: '3rem' }}>
-					<h1>{label}</h1>
-					<Module.default {...rest} />
-				</Container>
-			</Article>
-		);
-	}
-}
-
-const Home = () => (
-	<Article>
-		<Container>
-			<h1>GEL Examples</h1>
-			<p>Click one of the examples on the left to view it.</p>
-		</Container>
-	</Article>
-);
-
-// ==============================
-// Styled components
-// ==============================
-
-const Article = props => (
-	<article
-		css={{
-			flex: 1,
-			overflowY: 'auto',
-			paddingBottom: '4rem',
-			paddingLeft: 240,
-			paddingTop: '1rem',
-		}}
-		{...props}
-	/>
-);
-
-const Sidebar = ({ components, brand, setBrand, parent = '' }) => {
-	const [searchValue, setSearchValue] = useState('');
-	const location = useLocation();
-	const primaryColor = BRANDS[brand].COLORS.primary;
-
-	let navItems = components;
-
-	if (location.pathname !== '/') {
-		navItems = navItems.filter(component => location.pathname.startsWith(`/${component.parent}`));
-	} else {
-		navItems = navItems.filter(component => component.landing);
-	}
-
-	if (searchValue.length) {
-		navItems = navItems.filter(p => p.label.toLowerCase().includes(searchValue.toLowerCase()));
-	}
-
-	return (
-		<div
-			css={{
-				backgroundColor: '#F4F5F7',
-				borderRight: '1px solid rgba(0, 0, 0, 0.075)',
-				display: 'flex',
-				flexDirection: 'column',
-				height: '100vh',
-				position: 'fixed',
-				width: 240,
-			}}
-		>
-			<NavLink
-				to="/"
-				css={{
-					color: 'inherit',
-					display: 'block',
-					fontWeight: 500,
-					fontSize: '1.25rem',
-					padding: '1.25rem',
-					textDecoration: 'none',
-
-					':hover, :focus': {
-						textDecoration: 'inherit',
-					},
-
-					':focus': {
-						outlineOffset: -3,
-					},
-				}}
-			>
-				GEL
-			</NavLink>
-			<input
-				onChange={e => setSearchValue(e.target.value)}
-				placeholder="Search..."
-				type="search"
-				value={searchValue}
-				css={{
-					background: 0,
-					borderWidth: '1px 0',
-					borderStyle: 'solid',
-					borderColor: 'rgba(0, 0, 0, 0.075)',
-					boxSizing: 'border-box',
-					fontSize: 'inherit',
-					padding: '0.625rem 1.25rem',
-					width: '100%',
-
-					':focus': {
-						background: 'rgba(0,0,0,0.04)',
-						outlineOffset: -3,
-					},
-				}}
-			/>
-
-			<nav
-				css={{
-					flex: 1,
-					overflowY: 'auto',
-				}}
-			>
-				<ul
-					css={{
-						listStyle: 'none',
-						margin: '1rem 0',
-						padding: 0,
-					}}
-				>
-					{navItems.map(({ label, slug }) => (
-						<li data-test-nav key={slug} data-test-nav-link>
-							<NavLink
-								to={`/${slug}`}
-								css={{
-									backgroundColor: '#F4F5F7',
-									borderLeft: '3px solid transparent',
-									color: primaryColor,
-									display: 'block',
-									fontWeight: 500,
-									padding: '0.625rem 1.0625rem',
-									fontSize: '1rem',
-									textDecoration: 'none',
-
-									':hover, :focus': {
-										background: '#fafbfc',
-										textDecoration: 'inherit',
-									},
-
-									':focus': {
-										outlineOffset: -3,
-									},
-
-									'&.active': {
-										borderLeftColor: primaryColor,
-										color: 'inherit',
-									},
-								}}
-							>
-								{label}
-							</NavLink>
-						</li>
-					))}
-				</ul>
-			</nav>
-
-			<div
-				css={{
-					display: 'flex',
-					fontSize: '0.8125rem',
-				}}
-			>
-				{Object.keys(BRANDS).map(b => {
-					const isChecked = brand === b;
-					return (
-						<label
-							key={b}
-							css={{
-								alignItems: 'center',
-								borderTop: '1px solid',
-								borderTopColor: isChecked ? '#1F252C' : 'rgba(0, 0, 0, 0.1)',
-								boxSizing: 'border-box',
-								color: isChecked ? 'inherit' : '#1F252C',
-								cursor: 'pointer',
-								flex: 1,
-								fontWeight: 500,
-								justifyContent: 'center',
-								paddingBottom: '0.75rem',
-								paddingTop: '0.75rem',
-								textAlign: 'center',
-
-								input: {
-									height: 1,
-									position: 'absolute',
-									visibility: 'hidden',
-									width: 1,
-								},
-							}}
-						>
-							<input
-								name="brand"
-								type="radio"
-								onChange={e => setBrand(b)}
-								value={b}
-								checked={isChecked}
-							/>
-							{b}
-						</label>
-					);
-				})}
-			</div>
-		</div>
-	);
-};
-
-// ==============================
-// Render to node
-// ==============================
+import { AppDocs } from './components/AppDocs';
 
 export default components => {
 	const rootElement = document.getElementById('root');
-	ReactDOM.render(<App components={components} />, rootElement);
+	ReactDOM.render(<AppDocs components={components} />, rootElement);
 };
